backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
tikapi
fastapi
uvicorn
google-genai
google-cloud-bigquery
python-dotenv
tqdm
Jinja2
python-multipart
faiss-cpu